rename: modal components folder name
